# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SoulMatch AI - AI-powered matrimony platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{userId}:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get a specific user's public profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Public profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicProfile"
...
  /matches:
    get:
      operationId: getMatches
      tags: [matches]
      summary: Get curated matches for the current user
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Matches list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchesPage"
...
  /reports:
    post:
      operationId: createReport
      tags: [reports]
      summary: Report a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Report submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /admin/users:
    get:
      operationId: adminGetUsers
      tags: [admin]
      summary: Get all users with filters
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: role
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersPage"
  /admin/users/{userId}:
    get:
      operationId: adminGetUser
      tags: [admin]
      summary: Get single user details
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetail"
    patch:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Update user status/role
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetail"
  /admin/reports:
    get:
      operationId: adminGetReports
      tags: [admin]
      summary: Get all user reports
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, reviewed, resolved, dismissed]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Reports
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminReportsPage"
  /admin/reports/{reportId}:
    patch:
      operationId: adminResolveReport
      tags: [admin]
      summary: Resolve or dismiss a report
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportResolution"
      responses:
        "200":
          description: Resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /admin/verifications/{verificationId}:
    patch:
      operationId: adminProcessVerification
      tags: [admin]
      summary: Approve or reject a verification
      parameters:
        - name: verificationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerificationDecision"
      responses:
        "200":
          description: Processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── ANALYTICS ─────────────────────────────────────────────────────────────
...
  /analytics/users:
    get:
      operationId: getUserAnalytics
      tags: [analytics]
      summary: User growth and registration trends
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d, 1y]
            default: 30d
      responses:
        "200":
          description: User analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserAnalytics"
  /analytics/revenue:
    get:
      operationId: getRevenueAnalytics
      tags: [analytics]
      summary: Revenue and subscription trends
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d, 1y]
            default: 30d
      responses:
        "200":
          description: Revenue analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueAnalytics"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ── AUTH SCHEMAS ────────────────────────────────────────────────────────
...
    Photo:
      type: object
      required: [id, url, isPrimary]
      properties:
        id:
          type: integer
        url:
          type: string
        isPrimary:
          type: boolean
        publicId:
          type: ["string", "null"]
...
    PublicProfile:
      type: object
      required: [id, firstName, age, city]
      properties:
        id:
          type: integer
        firstName:
          type: string
        displayName:
          type: ["string", "null"]
        age:
          type: ["integer", "null"]
        occupation:
          type: ["string", "null"]
        education:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        religion:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        photos:
          type: array
          items:
            $ref: "#/components/schemas/Photo"
        verificationStatus:
          type: string
        isPremium:
          type: boolean
        compatibilityScore:
          type: ["integer", "null"]
        journeyProgress:
          type: integer
...
    MatchCard:
      type: object
      required: [userId, profile, compatibilityScore]
      properties:
        userId:
          type: integer
        profile:
          $ref: "#/components/schemas/PublicProfile"
        compatibilityScore:
          type: integer
        commonTraits:
          type: array
          items:
            type: string
        aiInsight:
          type: ["string", "null"]
        isNew:
          type: boolean
        isMutualInterest:
          type: boolean
    MatchesPage:
      type: object
      required: [matches, total, page, totalPages]
      properties:
        matches:
          type: array
          items:
            $ref: "#/components/schemas/MatchCard"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
...
    ReportInput:
      type: object
      required: [reportedUserId, reason]
      properties:
        reportedUserId:
          type: integer
        reason:
          type: string
          enum: [fake_profile, harassment, inappropriate_content, spam, scam, other]
        description:
          type: string
...
    AdminUsersPage:
      type: object
      required: [users, total, page, totalPages]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUserDetail"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
    AdminUserDetail:
      type: object
      required: [id, email, firstName, lastName, role, status, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
        status:
          type: string
          enum: [active, suspended, banned]
        verificationStatus:
          type: string
        isPremium:
          type: boolean
        subscriptionPlan:
          type: ["string", "null"]
        reportCount:
          type: integer
        lastActive:
          type: ["string", "null"]
        createdAt:
          type: string
        journeyProgress:
          type: integer
    AdminUserUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [user, premium, admin, superadmin]
        status:
          type: string
          enum: [active, suspended, banned]
    AdminReportsPage:
      type: object
      required: [reports, total, page, totalPages]
      properties:
        reports:
          type: array
          items:
            $ref: "#/components/schemas/AdminReport"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
    AdminReport:
      type: object
      required: [id, reason, status, createdAt]
      properties:
        id:
          type: integer
        reporterUserId:
          type: integer
        reportedUserId:
          type: integer
        reason:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, reviewed, resolved, dismissed]
        resolution:
          type: ["string", "null"]
        reporter:
          $ref: "#/components/schemas/PublicProfile"
        reported:
          $ref: "#/components/schemas/PublicProfile"
        createdAt:
          type: string
    ReportResolution:
      type: object
      required: [status, resolution]
      properties:
        status:
          type: string
          enum: [resolved, dismissed]
        resolution:
          type: string
        actionTaken:
          type: string
          enum: [none, warning, suspend, ban]
...
    VerificationDecision:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [approved, rejected]
        rejectionReason:
          type: string

    # ── ANALYTICS SCHEMAS ───────────────────────────────────────────────────
...
    UserAnalytics:
      type: object
      required: [period, data, totalGrowth]
      properties:
        period:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/TimeSeriesPoint"
        totalGrowth:
          type: number
        genderBreakdown:
          type: object
          properties:
            male:
              type: integer
            female:
              type: integer
            other:
              type: integer
        topLocations:
          type: array
          items:
            $ref: "#/components/schemas/LocationCount"
    TimeSeriesPoint:
      type: object
      required: [date, value]
      properties:
        date:
          type: string
        value:
          type: integer
    LocationCount:
      type: object
      required: [location, count]
      properties:
        location:
          type: string
        count:
          type: integer
    RevenueAnalytics:
      type: object
      required: [period, data, totalRevenue, planBreakdown]
      properties:
        period:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/TimeSeriesPoint"
        totalRevenue:
          type: number
        planBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/PlanRevenue"
    PlanRevenue:
      type: object
      required: [planId, planName, subscribers, revenue]
      properties:
        planId:
          type: string
        planName:
          type: string
        subscribers:
          type: integer
        revenue:
          type: number
...